Don't show query genes for queried data types, cytoband filter in Enrichments
