+++ src/pages/resultsView/enrichments/styles.module.scss
@@ -30,7 +30,7 @@
 
 .TableContainer {
     :global(.tableSearchInput) {
-        width: 196px !important;
+        width: 194px !important;
     }
 }
